fix(mobile): break react-native import chain for tests + push-only CI
Extract ClimbFilters/DEFAULT_FILTERS to climb-filter-types.ts and
getFilterKey to filter-key.ts so test files avoid the react-native
transitive import that Rolldown cannot parse (Flow syntax).

Fix filter-summary to skip grade bounds when grades is undefined
rather than showing uninformative "#N+" fallbacks.

Remove pull_request trigger from mobile-tests.yml so CI only fires
on push, not once per push and once per PR open/sync.
