+++ .github/workflows/mobile-tests.yml
@@ -10,14 +10,6 @@ on:
       - 'packages/shared-schema/**'
       - 'vite.config.ts'
       - '.github/workflows/mobile-tests.yml'
-  pull_request:
-    paths:
-      - 'packages/mobile/**'
-      - 'packages/shared/**'
-      - 'packages/board-constants/**'
-      - 'packages/shared-schema/**'
-      - 'vite.config.ts'
-      - '.github/workflows/mobile-tests.yml'
   workflow_dispatch:
 
 jobs:
