+++ packages/mobile/src/components/ClimbFilterSheet.tsx
@@ -16,20 +16,10 @@ import { springs } from '../theme/animations';
 import { brandColors } from '../theme/colors';
 import { iosSystemColors } from '../theme/ios-colors';
 import { spacing } from '../theme/tokens';
-
-export type ClimbFilters = {
-  minGrade?: number;
-  maxGrade?: number;
-  minAscents?: number;
-  minRating?: number;
-  sortBy: string;
-  sortOrder: string;
-};
-
-export const DEFAULT_FILTERS: ClimbFilters = {
-  sortBy: 'popular',
-  sortOrder: 'desc',
-};
+export type { ClimbFilters } from '../lib/climb-filter-types';
+export { DEFAULT_FILTERS } from '../lib/climb-filter-types';
+import type { ClimbFilters } from '../lib/climb-filter-types';
+import { DEFAULT_FILTERS } from '../lib/climb-filter-types';
 
 type ClimbFilterSheetProps = {
   visible: boolean;
